requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==5.0
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
drf-yasg==1.21.7
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
flake8==7.0.0
httpie==3.2.2
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.2.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
